Convert iOS integration to CocoaPod
This moves the iOS integration over to being a cocoapod modeled on the `react-native-Fabric` cocoapod.
This has the advantage of removing the necessary step of adding NewRelic manually to the `Podfile` and also of clearing up the framework linking phase to avoid issues we were seeing where `Foundation` wouldn't build.